requirements.txt
bibtexparser==1.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
future==0.18.2
greenlet==1.1.1
idna==3.2
msgpack==1.0.2
pynvim==0.4.3
pyparsing==2.4.7
python-slugify==5.0.2
requests==2.26.0
text-unidecode==1.3
urllib3==1.26.6
